+++ bin/publish.mli
@@ -14,22 +14,18 @@ val publish :
   ?publish_msg:string ->
   ?token:string Dune_release.Config.Cli.t ->
   ?dev_repo:string ->
-  pkg_names:string list ->
   version:Dune_release.Version.t option ->
   tag:Dune_release.Vcs.Tag.t option ->
   keep_v:bool Dune_release.Config.Cli.t ->
   dry_run:bool ->
-  publish_artefacts:[ `Distrib | `Doc ] list ->
   yes:bool ->
   draft:bool ->
   unit ->
   (int, Bos_setup.R.msg) result
 (** [publish ~build_dir ~opam ~change_log ~distrib_uri ~distrib_file
-    ~publish_msg ~name ~pkg_names ~version ~tag ~keep_v ~dry_run
-    ~publish_artefacts ~yes ~draft ()]
-    publishes the artefacts [publish_artefacts] of the package built with
-    [name], [version] and [tag]. Returns the exit code (0 for success, 1 for
-    failure) or error messages.
+    ~publish_msg ~version ~tag ~keep_v ~dry_run ~yes ~draft ()]
+    publishes the artefacts of the package built with [version] and [tag].
+    Returns the exit code (0 for success, 1 for failure) or error messages.
 
     - [keep_v] indicates whether the version is prefixed by 'v'. *)