+++ tests/bin/non-github-uri/run.t
@@ -53,8 +53,8 @@ archive should fail as only publishing to github is supported.
   dune-release: [ERROR] Github development repository URL could not be inferred
   [FAIL] lint of _build/whatever-0.1.0 and package whatever failure: 1 errors.
 
-(2) publish distrib
+(2) publish
 
-  $ dune-release publish distrib --dry-run > /dev/null
+  $ dune-release publish --dry-run > /dev/null
   dune-release: [ERROR] Github development repository URL could not be inferred from opam files.
   [3]